+++ passguard/src/main/java/com/example/passguard/models/User.java
@@ -1,6 +1,6 @@
 package com.example.passguard.models;
 
-public record User(int id, String email, String username) {
+public record User(long id, String email, String username) {
 
 
 }
